requirements.txt
opencv-python
numpy
pyyaml
insightface
onnxruntime
pillow
